requirements.txt
transformers
torch
plotly
streamli